Break util down to submodules
