+++ spotipy/util/credentials.py
@@ -1,64 +1,4 @@
-"""
-util
-====
-
-Utility module for your convenience <3
-
-The main motivation for this module is to make authorisation effortless.
-That goal is achieved by three mechanisms:
-reading configuration, an automatically refreshing access token and
-functions that implement everything needed to retrieve tokens.
-The effect is easy configuration, user authorisation and
-a strong independent token.
-
-.. code:: python
-
-    from spotipy import util
-
-    conf = util.credentials_from_environment()
-    app_token = util.request_client_token(*conf[:2])
-    user_token = util.prompt_for_user_token(*conf)
-
-    # Save the refresh token to avoid authenticating again
-    refresh_token = ...     # Load refresh token
-    user_token = util.refresh_user_token(*conf[:2], refresh_token)
-
-If you authenticate with a server but would still like to use
-:class:`RefreshingToken`, you can use the :class:`RefreshingCredentials`
-manager that is used by the functions above to create refreshing tokens.
-
-.. code:: python
-
-    cred = util.RefreshingCredentials(*conf)
-
-    # Client credentials flow
-    app_token = cred.request_client_token()
-
-    # Authorisation code flow
-    url = cred.user_authorisation_url()
-    code = ...  # Redirect user to login and retrieve code
-    user_token = cred.request_user_token(code)
-
-    # Reload a token
-    user_token = cred.refresh_user_token(refresh_token)
-
-Reading configuration from INI files is also possible.
-
-.. code:: python
-
-    cred = util.credentials_from_configfile(filename)
-
-This module exists solely to make developing applications easier.
-Some applications might have different needs,
-so please do create your own versions of these routines.
-Particularly, ``prompt_for_user_token`` opens up a web browser
-for the user to log in with, doesn't expose all parameters that are available
-when using lower-level functions, and forces the login dialog to be shown.
-"""
-
-import os
 import webbrowser
-import configparser
 
 from urllib.parse import urlparse, parse_qs
 from spotipy.auth import AccessToken, Token, Credentials
@@ -232,111 +172,6 @@ class RefreshingCredentials:
         """
         token = self._client.refresh_user_token(refresh_token)
         return RefreshingToken(token, self._client)
-
-
-def read_environment(*variables: str) -> tuple:
-    """
-    Read environment variables.
-
-    Parameters
-    ----------
-    variables
-        environment variable names to read
-
-    Returns
-    -------
-    tuple
-        variable values
-    """
-    return tuple(os.getenv(var, None) for var in variables)
-
-
-def credentials_from_environment(
-        client_id_var: str = 'SPOTIPY_CLIENT_ID',
-        client_secret_var: str = 'SPOTIPY_CLIENT_SECRET',
-        redirect_uri_var: str = 'SPOTIPY_REDIRECT_URI'
-) -> (str, str, str):
-    """
-    Read credentials from environment variables.
-
-    Parameters
-    ----------
-    client_id_var
-        name of the variable containing a client ID
-    client_secret_var
-        name of the variable containing a client secret
-    redirect_uri_var
-        name of the variable containing a redirect URI
-
-    Returns
-    -------
-    tuple
-        (client ID, client secret, redirect URI), None if not found
-    """
-    return read_environment(client_id_var, client_secret_var, redirect_uri_var)
-
-
-def read_configfile(*variables: str, config: dict) -> tuple:
-    """
-    Read variables from a config dict.
-
-    Parameters
-    ----------
-    variables
-        variable names to read
-    config
-        parsed config dict
-
-    Returns
-    -------
-    tuple
-        variable values
-    """
-    return tuple(config.get(var, None) for var in variables)
-
-
-def credentials_from_configfile(
-        file_path: str,
-        section: str = 'DEFAULT',
-        client_id_var: str = 'SPOTIPY_CLIENT_ID',
-        client_secret_var: str = 'SPOTIPY_CLIENT_SECRET',
-        redirect_uri_var: str = 'SPOTIPY_REDIRECT_URI'
-) -> (str, str, str):
-    """
-    Read credentials from a config file.
-
-    The configuration must be in INI format
-    as accepted by :class:`configparser.ConfigParser`.
-
-    Parameters
-    ----------
-    file_path
-        path of the file containing the credential variables
-    section
-        name of the section to read variables from
-    client_id_var
-        name of the variable containing a client ID
-    client_secret_var
-        name of the variable containing a client secret
-    redirect_uri_var
-        name of the variable containing a redirect URI
-
-    Returns
-    -------
-    tuple
-        (client ID, client secret, redirect URI), None if not found
-    """
-    c = configparser.ConfigParser()
-    c.optionxform = str
-    with open(file_path, 'r') as f:
-        c.read_file(f)
-
-    return read_configfile(
-        client_id_var,
-        client_secret_var,
-        redirect_uri_var,
-        config=c[section]
-    )
 
 
 def parse_code_from_url(url: str) -> str:
